requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
pytest

